Code refactoring New classes Facility and Term for GetMapping endpoints Mappers for Facility and Term classes New Controller's Handling for custom exception VaccinationDao class added

+++ src/main/java/com/dunin/medicalvaccinatesystem/dao/vaccination/repo/VaccinatedUserRepo.java
@@ -3,5 +3,5 @@ package com.dunin.medicalvaccinatesystem.dao.vaccination.repo;
 import com.dunin.medicalvaccinatesystem.dao.vaccination.model.VaccinatedUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VaccinatedUserRepo extends JpaRepository<VaccinatedUser, Integer> {
+public interface VaccinatedUserRepo extends JpaRepository<VaccinatedUser, Long> {
 }
