+++ src/main/java/com/dunin/medicalvaccinatesystem/dao/vaccination/repo/VaccinationFacilityRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VaccinationFacilityRepo extends JpaRepository<VaccinationFacility, Integer> {
+public interface VaccinationFacilityRepo extends JpaRepository<VaccinationFacility, Long> {
 }
